requirements.txt
camelot-py
pdftotext
unidecode
pandas
